requirements.txt
Pillow==10.0.0
Pillow==10.1.0
pipreqs==0.4.13
pyinstaller==5.6.1
